separate mocha node.js definition
